+++ next.clayui.com/content/docs/components/select.mdx
@@ -1,8 +1,8 @@
 ---
 title: "Select"
 description: "A form control element used to select from several provided options and enter data."
-packageNpm: "@clayui/select"
-packageVersion: "3.0.0-milestone.3"
+packageNpm: "@clayui/form"
+packageVersion: "3.0.0"
 packageStatus: "Beta"
 ---
