+++ packages/clay-form/src/index.tsx
@@ -9,16 +9,20 @@ import ClayForm from './Form';
 import ClayInput from './Input';
 import ClayRadio from './Radio';
 import ClayRadioGroup from './RadioGroup';
+import ClaySelect from './Select';
+import ClaySelectWithOption from './SelectWithOption';
 import {ClayInputWithAutocomplete} from './InputWithAutocomplete';
 import {ClayMultiSelectInput} from './MultiSelectInput';
 
 export {
-	ClayInputWithAutocomplete,
-	ClayMultiSelectInput,
 	ClayCheckbox,
 	ClayInput,
+	ClayInputWithAutocomplete,
+	ClayMultiSelectInput,
 	ClayRadio,
 	ClayRadioGroup,
+	ClaySelect,
+	ClaySelectWithOption,
 };
 
 export default ClayForm;
