+++ packages/clay-form/stories/index.tsx
@@ -12,6 +12,8 @@ import ClayForm, {
 	ClayMultiSelectInput,
 	ClayRadio,
 	ClayRadioGroup,
+	ClaySelect,
+	ClaySelectWithOption,
 } from '../src';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
@@ -219,3 +221,51 @@ const RadioGroupWithState = ({inline}: {inline?: boolean}) => {
 storiesOf('Components|ClayRadioGroup', module)
 	.add('default', () => <RadioGroupWithState />)
 	.add('inline', () => <RadioGroupWithState inline />);
+
+const options = [
+	{
+		label: 'Option 1',
+		value: '1',
+	},
+	{
+		label: 'Option 2',
+		value: '2',
+	},
+];
+
+storiesOf('Components|ClaySelect', module)
+	.add('with low-level APIs (composition)', () => (
+		<div className="sheet">
+			<div className="form-group">
+				<label htmlFor="mySelectId">{'Select Label'}</label>
+				<ClaySelect
+					aria-label="Select Label"
+					disabled={boolean('Disabled', false)}
+					id="mySelectId"
+					multiple={boolean('Multiple', false)}
+				>
+					{options.map(item => (
+						<ClaySelect.Option
+							key={item.value}
+							label={item.label}
+							value={item.value}
+						/>
+					))}
+				</ClaySelect>
+			</div>
+		</div>
+	))
+	.add('with high-level', () => (
+		<div className="sheet">
+			<div className="form-group">
+				<label htmlFor="mySelectId">{'Select Label'}</label>
+				<ClaySelectWithOption
+					aria-label="Select Label"
+					disabled={boolean('Disabled', false)}
+					id="mySelectId"
+					multiple={boolean('Multiple', false)}
+					options={options}
+				/>
+			</div>
+		</div>
+	));
