--- packages/clay-select/README.md
@@ -1,17 +1,0 @@
-# clay-select
-
-ClaySelect Component
-
-## Setup
-
-1. Install `yarn`
-
-2. Install local dependencies:
-
-```
-yarn
-```
-
-## Contribute
-
-We'd love to get contributions from you! Please, check our [Contributing Guidelines](https://github.com/liferay/clay/blob/master/CONTRIBUTING.md) to see how you can help us improve.
