--- packages/clay-select/package.json
@@ -1,39 +1,0 @@
-{
-	"name": "@clayui/select",
-	"version": "3.0.0-alpha.1",
-	"description": "ClaySelect component",
-	"license": "BSD-3-Clause",
-	"repository": "https://github.com/liferay/clay/tree/master/packages/clay-select",
-	"engines": {
-		"node": ">=0.12.0",
-		"npm": ">=3.0.0"
-	},
-	"main": "lib/index.js",
-	"types": "lib/index.d.ts",
-	"ts:main": "src/index.tsx",
-	"files": [
-		"lib",
-		"src"
-	],
-	"scripts": {
-		"build": "cross-env NODE_ENV=production babel src --root-mode upward --out-dir lib --extensions .ts,.tsx",
-		"build:types": "cross-env NODE_ENV=production tsc --project ./tsconfig.declarations.json",
-		"prepublishOnly": "yarn build && yarn build:types",
-		"test": "jest --config ../../jest.config.js"
-	},
-	"keywords": [
-		"clay",
-		"react"
-	],
-	"dependencies": {
-		"classnames": "^2.2.6"
-	},
-	"peerDependencies": {
-		"@clayui/css": "3.x",
-		"react": "^16.8.1",
-		"react-dom": "^16.8.1"
-	},
-	"browserslist": [
-		"extends browserslist-config-clay"
-	]
-}
