--- packages/clay-select/src/index.tsx
@@ -1,11 +1,0 @@
-/**
- * © 2019 Liferay, Inc. <https://liferay.com>
- *
- * SPDX-License-Identifier: BSD-3-Clause
- */
-
-import Select from './Select';
-import SelectWithOption from './SelectWithOption';
-
-export {SelectWithOption as ClaySelectWithOption};
-export default Select;
