--- packages/clay-select/stories/index.tsx
@@ -1,59 +1,0 @@
-/**
- * © 2019 Liferay, Inc. <https://liferay.com>
- *
- * SPDX-License-Identifier: BSD-3-Clause
- */
-import ClaySelect, {ClaySelectWithOption} from '../src';
-import React from 'react';
-import {boolean} from '@storybook/addon-knobs';
-import {storiesOf} from '@storybook/react';
-
-import '@clayui/css/lib/css/atlas.css';
-
-const options = [
-	{
-		label: 'Option 1',
-		value: '1',
-	},
-	{
-		label: 'Option 2',
-		value: '2',
-	},
-];
-
-storiesOf('Components|ClaySelect', module)
-	.add('with low-level APIs (composition)', () => (
-		<div className="sheet">
-			<div className="form-group">
-				<label htmlFor="mySelectId">{'Select Label'}</label>
-				<ClaySelect
-					aria-label="Select Label"
-					disabled={boolean('Disabled', false)}
-					id="mySelectId"
-					multiple={boolean('Multiple', false)}
-				>
-					{options.map(item => (
-						<ClaySelect.Option
-							key={item.value}
-							label={item.label}
-							value={item.value}
-						/>
-					))}
-				</ClaySelect>
-			</div>
-		</div>
-	))
-	.add('with high-level', () => (
-		<div className="sheet">
-			<div className="form-group">
-				<label htmlFor="mySelectId">{'Select Label'}</label>
-				<ClaySelectWithOption
-					aria-label="Select Label"
-					disabled={boolean('Disabled', false)}
-					id="mySelectId"
-					multiple={boolean('Multiple', false)}
-					options={options}
-				/>
-			</div>
-		</div>
-	));
